feat: generic EitherError type
