+++ src/errors/mod.rs
@@ -1,4 +1,5 @@
 pub mod admin_password_error;
+pub mod either_error;
 pub mod impl_response_error;
 pub mod insert_member_error;
 pub mod name_error;
