+++ src/util.rs
@@ -31,21 +31,3 @@ pub fn get_file_bytes(path: &str) -> Vec<u8> {
 
     buffer
 }
-
-#[macro_export]
-macro_rules! parsed_env {
-    ($name:literal, $type:ty) => {
-        $crate::expect_env!($name).parse::<$type>().expect(&format!(
-            "{} was not able to be parsed as {:?}",
-            $name,
-            stringify!($type)
-        ))
-    };
-}
-
-#[macro_export]
-macro_rules! expect_env {
-    ($name:literal) => {
-        std::env::var($name).expect(&format!("No {} environment variable found", $name))
-    };
-}
